rasterizer/pyproject.toml
[project]
name = "rasterizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
]
